test(ci): preserve stable shard layout for host circuit coverage
