--- .github/workflows/maint-1243-shard-stability.yml
@@ -1,42 +1,0 @@
-name: Preserve PR 1243 CI shard layout
-
-on:
-  push:
-    branches:
-      - maint/1039-host-circuit
-
-permissions:
-  contents: write
-
-jobs:
-  consolidate:
-    if: github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: oven-sh/setup-bun@v2
-        with:
-          bun-version: 1.3.14
-      - name: Consolidate new tests into existing suites
-        run: python3 scripts/maint_1243_shard_stability.py
-      - name: Remove standalone tests and maintenance files
-        run: git rm tests/upstream-host-circuit.test.ts tests/upstream-host-circuit-config.test.ts scripts/maint_1243_shard_stability.py .github/workflows/maint-1243-shard-stability.yml
-      - name: Install
-        run: bun install --frozen-lockfile
-      - name: Typecheck
-        run: bun run typecheck
-      - name: Focused tests
-        run: bun test tests/upstream-reachability.test.ts tests/config-user-edits.test.ts tests/responses-compaction-routing.test.ts
-      - name: Privacy scan
-        run: bun run privacy:scan
-      - name: Diff check
-        run: git diff --check
-      - name: Commit shard-stable tests
-        shell: bash
-        run: |
-          set -euxo pipefail
-          git config user.name "Wibias"
-          git config user.email "$(printf '%s%s%s' '37517432+Wibias' '@' 'users.noreply.github.com')"
-          git add tests/upstream-reachability.test.ts tests/config-user-edits.test.ts
-          git commit -m "test(ci): preserve stable shard layout for host circuit coverage"
-          git push origin HEAD:maint/1039-host-circuit
